Refactor

+++ src/main/java/com/matster/taxi/order/system/orderservice/repository/OrderRepository.java
@@ -1,7 +1,7 @@
 package com.matster.taxi.order.system.orderservice.repository;
 
 import com.matster.taxi.order.system.orderservice.entity.Order;
-import com.matster.taxi.order.system.orderservice.model.enums.OrderStatus;
+import com.matster.taxi.order.system.orderservice.enums.OrderStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,7 +11,7 @@ import java.util.Optional;
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
 
-    Optional<Order> findOrderById(Long id);
+    Optional<Order> findOrderById(Integer id);
 
     List<Order> findAllByOrderStatus(OrderStatus orderStatus);
 }
